# Modified HTML escaping to make it much faster