--- a/Cargo.toml
+++ b/Cargo.toml
@@ -25,6 +25,8 @@ serde = { version = "^0.8.0", optional = true }
 serde_json = { version = "^0.8.0", optional = true }
 serde_derive = { version = "^0.8.10", optional = true }
 pest = "^0.3.0"
+lazy_static = "0.2.1"
+regex = "0.1.77"
 
 [features]
 default = ["rustc_ser_type"]