@sunng87 2016-12-04
@Razican Thank you for this patch. I decided to use regex `replace_all`, which keeps code clean and easy to maintain. 

Implemented #118